CHORE(tests, matrix): added combined tests for matrix mult && inverse -> operation chainging resulted in precision issues, test marked with 'TODO' incase this becomes an issue later on

+++ srcs/ft_math/matrix/matrix_inverse.c
@@ -8,7 +8,7 @@ t_matrix	mtx_inverse(t_matrix m)
 	t_matrix	inv;
 	int			row;
 	int			col;
-	float		det;
+	double		det;
 
 	ft_assert(!eq_f(mtx_det(m), 0.0), __FILE__, __LINE__, "Error: matrix\
 			given to mtx_inverse() is not invertable (det == 0)");
